[IntroCSPy-week2] trying to make it easier to see the output and run modules
