+++ week2/solutions/char_is_in.py
@@ -9,5 +9,7 @@ def is_in(char, aStr):
     return is_in(char, aStr[0 : len(aStr)//2])
   return char == aStr[len(aStr)//2]
 
+print('is_in:')
 print('x is in fghklnz: ' + str(is_in('x', 'fghklnz')))
 print('n is in fghklnz: ' + str(is_in('n', 'fghklnz')))
+print('\n')
