+++ week2/solutions/check_odd.py
@@ -2,7 +2,9 @@ def odd(x):
   return bool(x % 2)
 
 
+print('odd:')
 x = 3
 print(str(x) + ' is odd: ' + str(odd(x)))
 y = 30
 print(str(y) + ' is odd: ' + str(odd(y)))
+print('\n')
