+++ week2/solutions/iter_power.py
@@ -14,3 +14,4 @@ print('using iter_power ' + str(base) + ' to the ' + str(exp) + ' power is ' + s
 base = 5
 exp = 5
 print('using iter_power ' + str(base) + ' to the ' + str(exp) + ' power is ' + str(iter_power(base, exp)))
+print('\n')
