+++ week2/solutions/recur_power.py
@@ -13,3 +13,4 @@ print('using recur_power ' + str(base) + ' to the ' + str(exp) + ' power is ' + 
 base = 5
 exp = 5
 print('using recur_power ' + str(base) + ' to the ' + str(exp) + ' power is ' + str(recur_power(base, exp)))
+print('\n')
