+++ week2/solutions/square_num.py
@@ -14,3 +14,4 @@ print('square:')
 print('5 squared is: ' + str(square(5)))
 print('fourth_power:')
 print('3 to the fourth power is ' + str(fourth_power(3)))
+print('\n')
